refactor: Order 관련 예외처리 코드 - common 패키지로 이동
