+++ src/main/kotlin/org/store/clothstar/common/config/exception/GlobalExceptionHandler.kt
@@ -13,6 +13,8 @@ import org.springframework.web.bind.annotation.RestControllerAdvice
 import org.store.clothstar.common.dto.ErrorResponseDTO
 import org.store.clothstar.common.dto.ValidErrorResponseDTO
 import org.store.clothstar.common.error.exception.*
+import org.store.clothstar.common.error.exception.order.InvalidOrderStatusException
+import org.store.clothstar.common.error.exception.order.OrderNotFoundException
 import java.util.function.Consumer
 
 @RestControllerAdvice
@@ -178,5 +180,18 @@ class GlobalExceptionHandler {
         )
 
         return ResponseEntity(errorResponseDTO, HttpStatus.INTERNAL_SERVER_ERROR)
+    }
+
+    // Order 관련 에러처리
+    @ExceptionHandler(OrderNotFoundException::class)
+    fun handleOrderNotFoundException(ex: OrderNotFoundException): ResponseEntity<ErrorResponseDTO> {
+        val errorResponseDTO = ErrorResponseDTO(ex.errorCode.status.value(), ex.errorCode.message)
+        return ResponseEntity(errorResponseDTO, ex.errorCode.status)
+    }
+
+    @ExceptionHandler(InvalidOrderStatusException::class)
+    fun handleInvalidOrderStatusException(ex: InvalidOrderStatusException): ResponseEntity<ErrorResponseDTO> {
+        val errorResponseDTO = ErrorResponseDTO(ex.errorCode.status.value(), ex.errorCode.message)
+        return ResponseEntity(errorResponseDTO, ex.errorCode.status)
     }
 }
