+++ src/main/kotlin/org/store/clothstar/common/dto/ErrorResponseDTO.kt
@@ -1,6 +1,6 @@
 package org.store.clothstar.common.dto
 
 class ErrorResponseDTO(
-    private val errorCode: Int,
-    private val message: String,
+    val errorCode: Int,
+    val message: String,
 )
