+++ src/main/kotlin/org/store/clothstar/common/error/exception/order/InvalidOrderStatusException.kt
@@ -1,0 +1,7 @@
+package org.store.clothstar.common.error.exception.order
+
+import org.store.clothstar.common.error.ErrorCode
+
+class InvalidOrderStatusException(
+    val errorCode: ErrorCode
+) : RuntimeException(errorCode.message)
