+++ src/main/kotlin/org/store/clothstar/common/error/exception/order/OrderNotFoundException.kt
@@ -1,0 +1,7 @@
+package org.store.clothstar.common.error.exception.order
+
+import org.store.clothstar.common.error.ErrorCode
+
+class OrderNotFoundException(
+    val errorCode: ErrorCode
+) : RuntimeException(errorCode.message)
