+++ src/main/kotlin/org/store/clothstar/order/controller/OrderSellerController.kt
@@ -12,8 +12,8 @@ import org.springframework.web.bind.annotation.PatchMapping
 import org.springframework.web.bind.annotation.PathVariable
 import org.springframework.web.bind.annotation.RequestMapping
 import org.springframework.web.bind.annotation.RestController
+import org.store.clothstar.common.dto.ErrorResponseDTO
 import org.store.clothstar.common.dto.MessageDTO
-import org.store.clothstar.order.exception.OrderErrorResponseDTO
 import org.store.clothstar.order.service.OrderSellerService
 
 @Tag(name = "OrderSeller", description = "판매자(OrderSeller)의 주문 정보 관리에 대한 API 입니다.")
@@ -28,9 +28,9 @@ class OrderSellerController(
         ApiResponse(responseCode = "200", description = "주문이 정상적으로 승인 되었습니다.",
             content = [Content(schema = Schema(implementation = MessageDTO::class))]),
         ApiResponse(responseCode = "400", description = "존재하지 않는 주문번호입니다.",
-            content = [Content(schema = Schema(implementation = OrderErrorResponseDTO::class))]),
+            content = [Content(schema = Schema(implementation = ErrorResponseDTO::class))]),
         ApiResponse(responseCode = "404", description = "주문이 승인대기 상태가 아니므로 요청을 처리할 수 없습니다.",
-            content = [Content(schema = Schema(implementation = OrderErrorResponseDTO::class))]),
+            content = [Content(schema = Schema(implementation = ErrorResponseDTO::class))]),
         ]
     )
     @PatchMapping("/{orderId}/approve")
@@ -45,9 +45,9 @@ class OrderSellerController(
         ApiResponse(responseCode = "200", description = "주문이 정상적으로 승인 되었습니다.",
             content = [Content(schema = Schema(implementation = MessageDTO::class))]),
         ApiResponse(responseCode = "400", description = "존재하지 않는 주문번호입니다.",
-            content = [Content(schema = Schema(implementation = OrderErrorResponseDTO::class))]),
+            content = [Content(schema = Schema(implementation = ErrorResponseDTO::class))]),
         ApiResponse(responseCode = "404", description = "주문이 승인대기 상태가 아니므로 요청을 처리할 수 없습니다.",
-            content = [Content(schema = Schema(implementation = OrderErrorResponseDTO::class))]),
+            content = [Content(schema = Schema(implementation = ErrorResponseDTO::class))]),
         ]
     )
     @PatchMapping("/{orderId}/cancel")
