--- src/main/kotlin/org/store/clothstar/order/exception/InvalidOrderStatusException.kt
@@ -1,4 +1,0 @@
-package org.store.clothstar.order.exception
-
-class InvalidOrderStatusException(val errorCode: OrderErrorCode) : RuntimeException(errorCode.message) {
-}
