--- src/main/kotlin/org/store/clothstar/order/exception/OrderErrorResponseDTO.kt
@@ -1,4 +1,0 @@
-package org.store.clothstar.order.exception
-
-class OrderErrorResponseDTO(val errorCode: Int, val message: String) {
-}
