--- src/main/kotlin/org/store/clothstar/order/exception/OrderNotFoundException.kt
@@ -1,4 +1,0 @@
-package org.store.clothstar.order.exception
-
-class OrderNotFoundException(val errorCode: OrderErrorCode) : RuntimeException(errorCode.message) {
-}
