+++ src/main/kotlin/org/store/clothstar/order/service/OrderSellerService.kt
@@ -3,9 +3,9 @@ package org.store.clothstar.order.service
 import org.springframework.data.repository.findByIdOrNull
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
+import org.store.clothstar.common.error.ErrorCode
 import org.store.clothstar.order.domain.vo.Status
-import org.store.clothstar.order.exception.OrderErrorCode
-import org.store.clothstar.order.exception.OrderNotFoundException
+import org.store.clothstar.common.error.exception.order.OrderNotFoundException
 import org.store.clothstar.order.repository.OrderRepository
 
 @Service
@@ -16,7 +16,7 @@ class OrderSellerService(
     @Transactional
     fun approveOrder(orderId: Long) {
         val order = orderRepository.findByIdOrNull(orderId)
-            ?: throw OrderNotFoundException(OrderErrorCode.NOT_FOUND_ORDER)
+            ?: throw OrderNotFoundException(ErrorCode.NOT_FOUND_ORDER)
         order.validateForStatusAndDeletedAt(Status.WAITING)
         order.updateStatus(Status.APPROVE)
     }
@@ -24,7 +24,7 @@ class OrderSellerService(
     @Transactional
     fun cancelOrder(orderId: Long) {
         val order = orderRepository.findByIdOrNull(orderId)
-            ?: throw OrderNotFoundException(OrderErrorCode.NOT_FOUND_ORDER)
+            ?: throw OrderNotFoundException(ErrorCode.NOT_FOUND_ORDER)
         order.validateForStatusAndDeletedAt(Status.WAITING)
         order.updateStatus(Status.CANCEL)
     }
