+++ src/test/kotlin/org/store/clothstar/order/service/OrderSellerServiceTest.kt
@@ -12,7 +12,7 @@ import org.junit.jupiter.api.extension.ExtendWith
 import org.springframework.data.repository.findByIdOrNull
 import org.store.clothstar.order.domain.Order
 import org.store.clothstar.order.domain.vo.Status
-import org.store.clothstar.order.exception.OrderNotFoundException
+import org.store.clothstar.common.error.exception.order.OrderNotFoundException
 import org.store.clothstar.order.repository.OrderRepository
 import kotlin.test.Test
